requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
jsonschema==4.17.3
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
protobuf==3.20.3
psycopg2==2.9.5
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.3
semver==3.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.8
srsly==2.4.6
streamlit==1.20.0
streamlit-javascript==0.1.5
thinc==8.1.9
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
yarl==1.8.2
zipp==3.15.0
